implements models
The user document will be created when someone logs in for the first time. And the messages will be created whenever a user sends a message.
Users can have the status of member, which allows them to see information about the message, like who wrote them and the date, and the status of admin, which allows them to delete messages.