implements member status
now the two main user types can be created: simple user and member.
Both types use the same form, with the difference that the member has to enter a secret member password (password check is not implemented yet).
Users can send messages but only members can see who wrote them or the date of creation (date is not implemented neither)