fixes bug in normal user signup
i had a bug where the member password would be validated even when signing up as a normal user, and that validation would always return an error in that case since the member password field renders conditionally.
So i made the member password validation conditional, only runs when the member password input is found in the form.

I know this is just a line but took some time to find this solution!