fixes bug at message send handler
since the message send handler was defined as an array (the standard way of using express-validator), it had to be redefined to an async function, and instead, the validation is run manually inside of it. Also, not including an else on the handler selector made the request go through both handlers when a message was sent, wich would get the program stuck.